Dev

See merge request observer/api!134

+++ observer/base/service/corpus.py
@@ -37,21 +37,37 @@ class CorpusAdd(Abstract):
     def add(self):
         
         keyword = getattr(self, 'keyword', '')
-        category_id = getattr(self, 'category_id', '')
-        industry_id = getattr(self, 'industry_id') if getattr(self, 'industry_id') else 0
+        category_id = getattr(self, 'category_id') if getattr(self, 'category_id', None) else '0006'
+        industry_id = getattr(self, 'industry_id') if getattr(self, 'industry_id', None) else -1
+        startTime = getattr(self, 'startTime', '0001-01-01')
+        endTime = getattr(self, 'endTime', '9999-12-30')
 
         if not category_id:
             return 400
 
-        if CorpusCategories.objects.filter(category_id=category_id, keyword=keyword, industry_id=industry_id).exists():
+        if CorpusCategories.objects.filter(category_id=category_id, keyword=keyword, industry_id=industry_id, user_id=self.user.id).exists():
             return 202
 
-        CorpusCategories(
-            keyword=keyword,
-            category_id=category_id,
-            industry_id=industry_id,
-            user_id=self.user.id,
-        ).save()
+        # 涉及到两个功能，语料添加和监测词添加，若不存在类别和行业则储存监测词
+        if getattr(self, 'category_id', None) or getattr(self, 'industry_id', None):
+            CorpusCategories(
+                keyword=keyword,
+                category_id=category_id,
+                industry_id=industry_id,
+                startTime=startTime,
+                stopTime=endTime,
+                user_id=self.user.id,
+            ).save()
+        else:
+            CorpusCategories(
+                keyword=keyword,
+                status=1,
+                category_id=category_id,
+                industry_id=industry_id,
+                startTime=startTime,
+                stopTime=endTime,
+                user_id=self.user.id,
+            ).save()
 
         return 200
 
@@ -85,14 +101,26 @@ class CorpusDelete(Abstract):
         keywords = getattr(self, 'keyword', '')
         category_ids = getattr(self, 'category_id', '')
         industry_ids = getattr(self, 'industry_id', '')
-            
+        corpus_ids = getattr(self, 'corpus_id', '')
+
         for (ids, keyword, category_id, industry_id, statu) in zip(del_ids.split(","), keywords.split(","), category_ids.split(","), industry_ids.split(","), status.split(",")):
             if statu == '1':
-                CrawlerTask_category(keyword, category_id, industry_id).remove()
-                CorpusCategories.objects.filter(id=ids).delete()
+                CrawlerTask_category(keyword, category_id, industry_id, self.user.id, ids).remove()
+                statuUp = CorpusCategories.objects.get(id=ids)
+                statuUp.status = 2  
+                statuUp.save()
             else:
-                CorpusCategories.objects.filter(id=ids).delete()
-                    
+                if statu == '0':
+                    CorpusCategories.objects.filter(id=ids).delete()
+                if statu == '2':
+                    corCategory = CorpusCategories.objects.get(id=ids)
+                    if self.user != 1:
+                        corCategory.user_id = 1
+                        corCategory.save()
+                    else:
+                        corCategory.delete()
+
+
         return 200
 
 
@@ -111,7 +139,7 @@ class CrawlerData(Abstract):
         industry_ids = getattr(self, 'industry_id')
 
         for (ids, keyword, category_id, industry_id) in zip(edit_ids.split(","), keywords.split(","), category_ids.split(","), industry_ids.split(",")):
-            CrawlerTask_category(keyword, category_id, industry_id, self.user.id).build()
+            CrawlerTask_category(keyword, category_id, industry_id, self.user.id, ids).build()
             corpus_categories = CorpusCategories.objects.get(id=ids)
             corpus_categories.status = status
             corpus_categories.save()
@@ -129,7 +157,7 @@ class CategoryListData(Abstract):
         cond = {
             'level': getattr(self, 'level', None),
             'parent': getattr(self, 'parent', None),
-            'name__icontains': getattr(self, 'text', None),
+            'name__icontains': getattr(self, 'name', None),
         }
 
         args = dict([k,v] for k, v in cond.items() if v)
